[BE] hotfix: Double → BigDecimal 로 수정

See merge request s13-fintech-finance-sub1/S13P21D108!75

+++ backend/moyeo/src/main/java/com/mo/moyeo/domain/exchange/reservation/dto/ExchangeReserveDto.java
@@ -3,14 +3,15 @@ package com.mo.moyeo.domain.exchange.reservation.dto;
 
 import com.mo.moyeo.domain.currency.entity.CurrencyType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ExchangeReserveDto (
         Long boxId,
         CurrencyType fromCurrency,
         CurrencyType toCurrency,
-        double amount,
-        double targetRate,
+        BigDecimal amount,
+        BigDecimal targetRate,
         LocalDate expiresAt
 ){
 }
