+++ backend/moyeo/src/main/java/com/mo/moyeo/domain/transaction/bank/dto/BankTransferDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 @AllArgsConstructor
@@ -11,7 +13,7 @@ public class BankTransferDTO {
 
     private String depositAccountNo;
     private String depositTransactionSummary;
-    private Double transactionBalance;
+    private BigDecimal transactionBalance;
     private String withdrawalAccountNo;
     private String withdrawalTransactionSummary;
     private String userKey;
